tests: add integration tests
This updates the Grafana service in the docker-compose file to have some
dashboards and datasources provisioned, then adds some integration tests
for each tool to the test suite. These can be run by adding the
`--integration-tests` flag to the `pytest` invocation.

There are also tests for cloud-only functionality (i.e. Grafana Incident)
which can be run by adding the `--cloud-only` flag to the `pytest`
invocation.